GCRApps/toto/dodo/requirements.txt
Flask==2.0.1
dataclasses==0.6
requests==2.26.0
beautifulsoup4==4.8.0
protobuf==3.17.3
python_dateutil==2.8.2
google-cloud-bigquery==2.26.0
gunicorn==20.1.0
google-cloud-aiplatform==1.7.0
flask_wtf==1.0.1
flask-bootstrap==3.3.7.1
wtforms==3.0.1
